+++ cli/ChangePasswordCommand.php
@@ -12,6 +12,7 @@ use Grav\Console\ConsoleCommand;
 use Grav\Common\Grav;
 use Grav\Common\File\CompiledYamlFile;
 use Grav\Common\User\User;
+use Grav\Plugin\Login\Login;
 use RocketTheme\Toolbox\ResourceLocator\UniformResourceLocator;
 use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Helper\Helper;
@@ -24,10 +25,11 @@ use Symfony\Component\Console\Question\Question;
  */
 class ChangePasswordCommand extends ConsoleCommand
 {
-    /**
-     * @var array
-     */
+    /** @var array */
     protected $options = [];
+
+    /** @var Login */
+    protected $login;
 
     /**
      * Configure the command
@@ -59,6 +61,14 @@ class ChangePasswordCommand extends ConsoleCommand
      */
     protected function serve()
     {
+        include __DIR__ . '/../vendor/autoload.php';
+
+        $grav = Grav::instance();
+        if (!isset($grav['login'])) {
+            $grav['login'] = new Login($grav);
+        }
+        $this->login = $grav['login'];
+
         $this->options = [
             'user'        => $this->input->getOption('user'),
             'password1'   => $this->input->getOption('password')
@@ -76,7 +86,13 @@ class ChangePasswordCommand extends ConsoleCommand
             // Get username and validate
             $question = new Question('Enter a <yellow>username</yellow>: ');
             $question->setValidator(function ($value) {
-                return $this->validate('user', $value);
+                $this->validate('user', $value);
+
+                if (!User::find($value, ['username'])->exists()) {
+                    throw new \RuntimeException('Username "' . $value . '" does not exist, please pick another username');
+                };
+
+                return true;
             });
 
             $username = $helper->ask($this->input, $this->output, $question);
@@ -111,7 +127,7 @@ class ChangePasswordCommand extends ConsoleCommand
         $oldUserFile = CompiledYamlFile::instance($locator->findResource('account://' . $username . YAML_EXT, true, true));
         $oldData = (array)$oldUserFile->content();
         
-        //Set the password feild to new password
+        //Set the password field to new password
         $oldData['password'] = $data['password'];
         
         // Create user object and save it using oldData (with updated password)
@@ -135,50 +151,15 @@ class ChangePasswordCommand extends ConsoleCommand
     }
 
     /**
-     * @param        $type
-     * @param        $value
+     * @param string $type
+     * @param mixed  $value
      * @param string $extra
      *
-     * @return mixed
+     * @return string
      */
     protected function validate($type, $value, $extra = '')
     {
-        /** @var Config $config */
-        $config = Grav::instance()['config'];
-
-        /** @var UniformResourceLocator $locator */
-        $locator = Grav::instance()['locator'];
-
-        $username_regex = '/' . $config->get('system.username_regex') . '/';
-        $pwd_regex      = '/' . $config->get('system.pwd_regex') . '/';
-
-        switch ($type) {
-            case 'user':
-                if (!preg_match($username_regex, $value)) {
-                    throw new \RuntimeException('Username should be between 3 and 16 characters, including lowercase letters, numbers, underscores, and hyphens. Uppercase letters, spaces, and special characters are not allowed');
-                }
-                if (!$locator->findResource('account://' . $value . YAML_EXT)) {
-                    throw new \RuntimeException('Username "' . $value . '" does not exist, please pick another username');
-                }
-
-                break;
-
-            case 'password1':
-                if (!preg_match($pwd_regex, $value)) {
-                    throw new \RuntimeException('Password must contain at least one number and one uppercase and lowercase letter, and at least 8 or more characters');
-                }
-
-                break;
-
-            case 'password2':
-                if (strcmp($value, $extra)) {
-                    throw new \RuntimeException('Passwords did not match.');
-                }
-
-                break;
-        }
-
-        return $value;
+        return $this->login->validateField($type, $value, $extra);
     }
 
     /**